Conflicts:
	www/app_services_view.py
	www/compose_service_view.py
	www/services_view.py
	www/templates/www/include/tab_deployed.html
	www/templates/www/include/tab_settings.html
	www/templates/www/service_my.html
	www/templates/www/team.html

+++ www/views/ajax/service_actions.py
@@ -889,6 +889,11 @@ class ServicePort(AuthedView):
         if action == 'change_protocol':
             protocol = request.POST.get("value")
             deal_port.protocol = protocol
+            # 判断stream协议的对外数量
+            if protocol == "stream":
+                outer_num = TenantServicesPort.objects.filter(service_id=self.service.service_id, protocol="stream").count()
+                if outer_num == 1:
+                    return JsonResponse({"success": False, "code": 410, "info": u"对外stream端口只支持一个"})
             data.update({"modified_field": "protocol", "current_value": protocol})
         elif action == 'open_outer':
             deal_port.is_outer_service = True
@@ -959,6 +964,12 @@ class ServicePort(AuthedView):
             if not success:
                 return JsonResponse({"success": False, "info": reason, "code": 400}, status=400)
             else:
+                if TenantServicesPort.objects.filter(service_id=self.service.service_id, container_port=deal_port.container_port, is_outer_service=True).count()>0:
+                    return JsonResponse({"success": False, "code": 400, "info": u"请关闭对外服务"}, status=400)
+            
+                if TenantServicesPort.objects.filter(service_id=self.service.service_id, container_port=deal_port.container_port, is_inner_service=True).count()>0:
+                    return JsonResponse({"success": False, "code": 400, "info": u"请关闭对内服务"}, status=400)
+
                 old_port = deal_port.container_port
                 deal_port.container_port = new_port
                 TenantServiceEnvVar.objects.filter(service_id=deal_port.service_id, container_port=old_port).update(container_port=new_port)
@@ -1110,7 +1121,11 @@ class ServiceNewPort(AuthedView):
 
             if TenantServicesPort.objects.filter(service_id=self.service.service_id, container_port=port_port).exists():
                 return JsonResponse({"success": False, "code": 409, "info": u"容器端口冲突"})
-
+            # 对外stream只能开一个
+            if port_protocol == "stream":
+                outer_num = TenantServicesPort.objects.filter(service_id=self.service.service_id, protocol="stream").count()
+                if outer_num == 1:
+                    return JsonResponse({"success": False, "code": 410, "info": u"对外stream端口只支持一个"})
             mapping_port = 0
             if port_inner == 1 and port_protocol == "stream":
                 mapping_port = baseService.prepare_mapping_port(self.service, port_port)
@@ -1131,6 +1146,14 @@ class ServiceNewPort(AuthedView):
             num = ServiceDomain.objects.filter(service_id=self.service.service_id, container_port=port_port).count()
             if num > 0:
                 return JsonResponse({"success": False, "code": 409, "info": u"请先解绑该端口绑定的域名"})
+            
+            if TenantServicesPort.objects.filter(service_id=self.service.service_id, container_port=port_port, is_outer_service=True).count()>0:
+                return JsonResponse({"success": False, "code": 409, "info": u"请关闭对外服务"})
+            
+            if TenantServicesPort.objects.filter(service_id=self.service.service_id, container_port=port_port, is_inner_service=True).count()>0:
+                return JsonResponse({"success": False, "code": 409, "info": u"请关闭对内服务"})
+            
+            
             TenantServicesPort.objects.filter(service_id=self.service.service_id, container_port=port_port).delete()
             TenantServiceEnvVar.objects.filter(service_id=self.service.service_id, container_port=port_port).delete()
             ServiceDomain.objects.filter(service_id=self.service.service_id, container_port=port_port).delete()
@@ -1194,7 +1217,7 @@ class ServiceVolumeView(AuthedView):
                 volume_path = request.POST.get("volume_path")
                 # category = self.service.category
                 language = self.service.language
-                if language == "docker":
+                if self.service.image != "goodrain.me/runner":
                     if not volume_path.startswith("/"):
                         result["status"] = "failure"
                         result["code"] = "303"
@@ -1224,7 +1247,11 @@ class ServiceVolumeView(AuthedView):
                             result["status"] = "failure"
                             result["code"] = "306"
                             return JsonResponse(result)
-
+                
+                if self.service.host_path is None or self.service.host_path == "":
+                    self.service.host_path = "/grdata/tenant/" + self.service.tenant_id + "/service/" + self.service.service_id
+                    self.service.save()
+                    
                 volume_id = baseService.create_service_volume(self.service, volume_path)
                 if volume_id:
                     result["volume"] = {
